refactor: Split responsibilities inside sim between 3 subcrates

+++ src/core/types.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub struct AgentId(pub usize);
+pub struct AgentId(pub u64);
 
 #[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub struct CommunityId(pub usize);
+pub struct CommunityId(pub u64);
